+++ src/SharpTerminal/RemoteSignalDecoder.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Json;
+using System.ComponentModel;
+using System.Windows.Forms;
+
+namespace SharpTerminal
+{
+    
+        
+}